# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mappers Protocol API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  pending: 'pending',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Job {
  id: number;
  jobId: string;
  clientPubkey: string;
  freelancerPubkey: string;
  oraclePubkey: string;
  /** Amount in lamports as string (u64 safe) */
  amountLamports: string;
  status: JobStatus;
  /** @nullable */
  description: string | null;
  /**
     * JSON-encoded array of criteria strings
     * @nullable
     */
  acceptanceCriteria: string | null;
  /**
     * Most recent transaction signature
     * @nullable
     */
  txSig: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface JobInput {
  /** @maxLength 32 */
  jobId: string;
  clientPubkey: string;
  freelancerPubkey: string;
  oraclePubkey: string;
  amountLamports: string;
  description?: string;
  acceptanceCriteria?: string[];
}

export type JobUpdateStatus = typeof JobUpdateStatus[keyof typeof JobUpdateStatus];


export const JobUpdateStatus = {
  pending: 'pending',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface JobUpdate {
  status?: JobUpdateStatus;
  txSig?: string;
  description?: string;
}

export type DeliverableInputDeliverableType = typeof DeliverableInputDeliverableType[keyof typeof DeliverableInputDeliverableType];


export const DeliverableInputDeliverableType = {
  url: 'url',
  ipfs: 'ipfs',
  text: 'text',
  json: 'json',
} as const;

export interface DeliverableInput {
  description: string;
  acceptanceCriteria: string[];
  deliverable: string;
  deliverableType: DeliverableInputDeliverableType;
}

/**
 * @nullable
 */
export type SubmitResultOutcome = typeof SubmitResultOutcome[keyof typeof SubmitResultOutcome] | null;


export const SubmitResultOutcome = {
  RELEASE: 'RELEASE',
  REFUND: 'REFUND',
  ESCALATE: 'ESCALATE',
} as const;

export interface SubmitResult {
  success: boolean;
  jobId: string;
  /** @nullable */
  outcome: SubmitResultOutcome;
  /** @nullable */
  txSig: string | null;
  /** @nullable */
  error: string | null;
}

export interface DashboardStats {
  total: number;
  pending: number;
  completed: number;
  cancelled: number;
  /** Sum of pending escrow amounts as string */
  totalEscrowedLamports: string;
}

export interface OracleHealth {
  status: string;
  pendingJobs: number;
  timestamp: string;
}

export type ListJobsParams = {
status?: ListJobsStatus;
clientPubkey?: string;
};

export type ListJobsStatus = typeof ListJobsStatus[keyof typeof ListJobsStatus];


export const ListJobsStatus = {
  pending: 'pending',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;