# lib/api-zod/src/generated/types/deliverableInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mappers Protocol API
 * OpenAPI spec version: 0.1.0
 */
import type { DeliverableInputDeliverableType } from './deliverableInputDeliverableType';

export interface DeliverableInput {
  description: string;
  acceptanceCriteria: string[];
  deliverable: string;
  deliverableType: DeliverableInputDeliverableType;
}