# lib/api-zod/src/generated/types/oracleHealth.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Mappers Protocol API
 * OpenAPI spec version: 0.1.0
 */

export interface OracleHealth {
  status: string;
  pendingJobs: number;
  timestamp: string;
}